+++ android/src/main/java/me/carda/awesome_notifications/Definitions.java
@@ -67,6 +67,7 @@ public interface Definitions {
     String CHANNEL_METHOD_SHOW_NOTIFICATION_PAGE = "showNotificationPage";
     String CHANNEL_METHOD_IS_NOTIFICATION_ALLOWED = "isNotificationAllowed";
     String CHANNEL_METHOD_REQUEST_NOTIFICATIONS = "requestNotifications";
+    String CHANNEL_METHOD_CHECK_PERMISSIONS = "checkPermissions";
 
     String CHANNEL_METHOD_GET_BADGE_COUNT = "getBadgeCount";
     String CHANNEL_METHOD_SET_BADGE_COUNT = "setBadgeCount";
@@ -166,12 +167,15 @@ public interface Definitions {
     String NOTIFICATION_ENABLED = "enabled";
     String NOTIFICATION_AUTO_DISMISSABLE = "autoDismissable";
     String NOTIFICATION_IS_DANGEROUS_OPTION = "isDangerousOption";
+    String NOTIFICATION_PERMISSIONS = "permissions";
 
     String NOTIFICATION_SHOW_IN_COMPACT_VIEW = "showInCompactView";
     String NOTIFICATION_LOCKED = "locked";
     String NOTIFICATION_DISPLAY_ON_FOREGROUND = "displayOnForeground";
     String NOTIFICATION_DISPLAY_ON_BACKGROUND = "displayOnBackground";
     String NOTIFICATION_ICON = "icon";
+    String NOTIFICATION_WAKE_UP_SCREEN = "wakeUpScreen";
+    String NOTIFICATION_CRITICAL_ALERT = "criticalAlert";
     String NOTIFICATION_PLAY_SOUND = "playSound";
     String NOTIFICATION_SOUND_SOURCE = "soundSource";
     String NOTIFICATION_ENABLE_VIBRATION = "enableVibration";
@@ -241,5 +245,7 @@ public interface Definitions {
         put(Definitions.NOTIFICATION_ONLY_ALERT_ONCE, false);
         put(Definitions.NOTIFICATION_SHOW_IN_COMPACT_VIEW, true);
         put(Definitions.NOTIFICATION_IS_DANGEROUS_OPTION, false);
+        put(Definitions.NOTIFICATION_WAKE_UP_SCREEN, false);
+        put(Definitions.NOTIFICATION_CRITICAL_ALERT, false);
     }};
 }