+++ lib/src/awesome_notifications_core.dart
@@ -28,6 +28,8 @@ import 'package:awesome_notifications/src/utils/bitmap_utils.dart';
 import 'package:awesome_notifications/src/utils/date_utils.dart';
 import 'package:rxdart/rxdart.dart' show BehaviorSubject;
 
+import 'enumerators/notification_permission.dart';
+
 class AwesomeNotifications {
   static String? rootNativePath;
 
@@ -285,7 +287,7 @@ class AwesomeNotifications {
     await _channel.invokeMethod(CHANNEL_METHOD_SHOW_NOTIFICATION_PAGE);
   }
 
-  /// Check if the notifications are permitted
+  /// Check if the notifications are globally permitted
   Future<bool> isNotificationAllowed() async {
     final bool isAllowed =
         await _channel.invokeMethod(CHANNEL_METHOD_IS_NOTIFICATION_ALLOWED);
@@ -293,10 +295,48 @@ class AwesomeNotifications {
   }
 
   /// Prompts the user to enabled notifications
-  Future<bool> requestPermissionToSendNotifications() async {
-    final bool isAllowed =
-        await _channel.invokeMethod(CHANNEL_METHOD_REQUEST_NOTIFICATIONS);
+  Future<bool> requestPermissionToSendNotifications(
+      {List<NotificationPermission> permissions = const [
+        NotificationPermission.Badge,
+        NotificationPermission.Alert,
+        NotificationPermission.Sound
+      ]}) async {
+    final List<String> permissionList = [];
+    for (final permission in permissions) {
+      String? permissionValue = AssertUtils.toSimpleEnumString(permission);
+      if (permissionValue != null) permissionList.add(permissionValue);
+    }
+
+    final bool isAllowed = await _channel.invokeMethod(
+        CHANNEL_METHOD_REQUEST_NOTIFICATIONS, permissionList);
     return isAllowed;
+  }
+
+  /// Check each individual permission to send notifications and returns only the allowed permissions
+  Future<List<NotificationPermission>> checkPermissionList(
+      {List<NotificationPermission> permissions = const [
+        NotificationPermission.Badge,
+        NotificationPermission.Alert,
+        NotificationPermission.Sound
+      ]}) async {
+    List<String> permissionList = [];
+    for (final permission in permissions) {
+      String? permissionValue = AssertUtils.toSimpleEnumString(permission);
+      if (permissionValue != null) permissionList.add(permissionValue);
+    }
+
+    permissionList = await _channel.invokeMethod(
+        CHANNEL_METHOD_REQUEST_NOTIFICATIONS, permissionList);
+
+    List<NotificationPermission> allowed = [];
+    for (final permission in permissionList) {
+      NotificationPermission? permissionValue =
+          AssertUtils.enumToString<NotificationPermission>(
+              permission, NotificationPermission.values, null);
+      if (permissionValue != null) allowed.add(permissionValue);
+    }
+
+    return allowed;
   }
 
   /// List all active scheduled notifications.